feat: folder restructure to make the project more readable, testable and scalable, add clearscore section, add github actions
