+++ src/app/(routes)/clearscore/page.tsx
@@ -1,0 +1,3 @@
+import ClearScore from '../../_/components/clearscore';
+
+export default ClearScore;
