+++ src/app/(routes)/page.tsx
@@ -1,0 +1,3 @@
+import HomePage from '../_/components/home-page';
+
+export default HomePage;
