+++ src/app/_/components/clearscore/index.ts
@@ -1,0 +1,3 @@
+import ClearScore from "./clearscore";
+
+export default ClearScore
