+++ src/app/_/components/home-page/home-page.module.css
@@ -1,0 +1,286 @@
+.section {
+  margin-top: var(--spacing-64);
+}
+
+.section__title {
+  font-size: var(--font-size-20);
+  padding: 0 var(--spacing-16);
+  font-weight: 600;
+}
+
+.visually-hidden-item {
+  width: 0;
+  height: 0;
+  opacity: 0;
+  position: absolute;
+  inset: 0;
+}
+
+.list {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-12);
+
+  @media (min-width: 768px) {
+    gap: var(--spacing-24);
+  }
+}
+
+.list:not(:first-child) {
+  margin-top: var(--spacing-32);
+}
+
+.list__item {
+  position: relative;
+  display: flex;
+  gap: var(--spacing-24);
+  list-style: none;
+  font-weight: 300;
+  padding: 0 var(--spacing-16);
+}
+
+.list__item--card {
+  padding: 0;
+}
+
+.skill {
+  font-size: var(--font-size-14);
+
+  @media (min-width: 512px) {
+    font-size: var(--font-size-16);
+  }
+}
+
+.experience-card {
+  display: flex;
+  flex-direction: column;
+  padding: var(--spacing-16);
+  border-color: white;
+  width: 100%;
+
+  @media (min-width: 768px) {
+    flex-direction: row-reverse;
+    padding: var(--spacing-24) var(--spacing-16);
+  }
+}
+
+.experience-card:hover {
+  color: rgb(102 238 255);
+  border-color: rgb(102 238 255);
+}
+
+.experience-card:hover > .experience-card__content {
+  text-decoration: underline;
+}
+
+.experience-card:hover > .experience-card__content > .experience-card__content__icon {
+  transition: transform 0.25s;
+  fill: rgb(102 238 255);
+  transform: translate(var(--spacing-4), calc(-1 * var(--spacing-4)));
+}
+
+.experience-card__content {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--spacing-4);
+  font-weight: 600;
+  width: 100%;
+  fill: white;
+}
+
+.experience-card__content__icon {
+  width: 1rem;
+  margin-left: var(--spacing-8);
+}
+
+.experience-card__content__role {
+  width: 100%;
+}
+
+.experience-card__content__description {
+  margin-top: var(--spacing-8);
+  font-size: var(--font-size-12);
+  opacity: 0.7;
+  font-weight: 300;
+
+  @media (min-width: 512px) {
+    font-size: var(--font-size-14);
+  }
+}
+
+.experience-card__aside {
+  position: absolute;
+  right: var(--spacing-16);
+  top: var(--spacing-40);
+  font-size: var(--font-size-12);
+  width: 100%;
+  max-width: 30%;
+
+  @media (min-width: 768px) {
+    position: relative;
+    inset: 0;
+    font-size: var(--font-size-14);
+  }
+}
+
+.experience-card__aside__date {
+  display: flex;
+  align-items: center;
+  white-space: nowrap;
+  justify-content: flex-end;
+
+  @media (min-width: 768px) {
+    justify-content: flex-start;
+  }
+}
+
+.separator {
+  width: 2rem;
+  border: 0.5px solid;
+  border-color: inherit;
+  transform: rotate(145deg);
+  margin: 0 var(--spacing-4);
+}
+
+.experience-card__link {
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  border-radius: var(--border-radius-16);
+}
+
+.experience-card__link:hover {
+  background-color: rgb(0 255 255 / 20%);
+}
+
+.experience-card__link__label {
+  opacity: 0;
+}
+
+.project-card {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+
+  @media (min-width: 768px) {
+    flex-direction: row-reverse;
+  }
+}
+
+.project-card__content {
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  padding: var(--spacing-16);
+  fill: white;
+
+  @media (min-width: 768px) {
+    padding: var(--spacing-32) var(--spacing-16) var(--spacing-32) 0;
+  }
+}
+
+.project-card__content:hover {
+  color: rgb(102 238 255);
+  text-decoration: underline;
+  text-underline-position: from-font;
+}
+
+.project-card__content:hover > .project-card__content__header__icon {
+  transition: transform 0.25s;
+  fill: rgb(102 238 255);
+  transform: translate(var(--spacing-4), calc(-1 * var(--spacing-4)));
+}
+
+.project-card__content:hover::before {
+  content: '';
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  border-radius: var(--border-radius-16);
+  background-color: rgb(0 255 255 / 20%);
+  z-index: -1;
+}
+
+.project-card__content__header__icon {
+  width: 1rem;
+  fill: inherit;
+  margin-left: var(--spacing-8);
+}
+
+.project-card__content__description {
+  opacity: 0.7;
+  margin-top: var(--spacing-8);
+  font-size: var(--font-size-12);
+
+  @media (min-width: 512px) {
+    font-size: var(--font-size-14);
+  }
+}
+
+.project-card__content__link {
+  width: 100%;
+  max-width: 80%;
+  height: 100%;
+  position: absolute;
+  inset: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+
+  @media (min-width: 768px) {
+    justify-content: flex-start;
+    left: unset;
+    top: 0;
+    right: 0;
+    max-width: 70%;
+  }
+}
+
+.project-card__aside {
+  width: 100%;
+  max-width: 20%;
+
+  @media (min-width: 768px) {
+    max-width: 30%;
+  }
+}
+
+.project-card__aside__link {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: flex-end;
+  padding: var(--spacing-32) var(--spacing-16);
+
+  @media (min-width: 768px) {
+    justify-content: flex-start;
+  }
+}
+
+.project-card__aside__link:hover::before {
+  content: '';
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  border-radius: var(--border-radius-16);
+  background-color: rgb(0 255 255 / 20%);
+  z-index: -1;
+}
+
+.project-card__aside__link:hover > svg {
+  fill: rgb(102 238 255);
+}
+
+.project-card__aside__link__icon {
+  width: 1.5rem;
+  height: 1.5rem;
+  fill: white;
+
+  @media (min-width: 512px) {
+    width: 2rem;
+    height: 2rem;
+  }
+}
