+++ src/app/_/components/home-page/index.ts
@@ -1,0 +1,6 @@
+import HomePage from './home-page';
+
+import Layout from './layout';
+
+export { Layout };
+export default HomePage;
