--- src/app/page.module.css
@@ -1,413 +1,0 @@
-.root {
-  width: 100%;
-  height: 100%;
-  min-height: 100vh;
-  display: grid;
-  grid-template-rows: min-content min-content 1fr;
-  grid-template-areas: 'header' 'footer' 'main-content';
-  column-gap: 12rem;
-  row-gap: var(--spacing-40);
-  padding: var(--spacing-48) var(--spacing-16);
-  margin: 0 auto;
-  color: white;
-
-  @media (min-width: 1024px) {
-    grid-template-rows: 1fr 1fr;
-    grid-template-columns: 3fr 5fr;
-    grid-template-areas: 'header main-content' 'footer main-content';
-    max-width: 80rem;
-  }
-}
-
-.header-group {
-  position: relative;
-  grid-area: header;
-  max-width: 30rem;
-  height: fit-content;
-  display: flex;
-  flex-direction: column;
-  row-gap: var(--spacing-4);
-  padding: 0 var(--spacing-16);
-
-  @media (min-width: 512px) {
-    display: grid;
-    grid-template-columns: repeat(3, min-content);
-    align-items: center;
-  }
-
-  @media (min-width: 1024px) {
-    position: fixed;
-  }
-}
-
-.header {
-  font-size: var(--font-size-18);
-  font-weight: 700;
-  white-space: nowrap;
-  text-transform: uppercase;
-
-  @media (min-width: 512px) {
-    font-size: var(--font-size-24);
-  }
-}
-
-.separator {
-  width: 2rem;
-  border: 0.5px solid;
-  border-color: inherit;
-  transform: rotate(145deg);
-  margin: 0 var(--spacing-4);
-}
-
-.separator--header {
-  display: none;
-
-  @media (min-width: 512px) {
-    display: flex;
-    width: 4.5rem;
-  }
-}
-
-.subheader {
-  font-size: var(--font-size-18);
-  font-weight: 300;
-  justify-self: end;
-  white-space: nowrap;
-  text-transform: capitalize;
-
-  @media (min-width: 512px) {
-    font-size: var(--font-size-24);
-  }
-}
-
-.description {
-  grid-column: 1 / span 3;
-  font-weight: 300;
-  font-size: var(--font-size-12);
-  margin-top: var(--spacing-4);
-  opacity: 0.7;
-
-  @media (min-width: 512px) {
-    font-size: var(--font-size-16);
-  }
-}
-
-.main {
-  grid-area: main-content;
-  padding-top: var(--spacing-16);
-}
-
-.section {
-  margin-top: var(--spacing-64);
-}
-
-.section__title {
-  font-size: var(--font-size-20);
-  padding: 0 var(--spacing-16);
-  font-weight: 600;
-}
-
-.visually-hidden-item {
-  width: 0;
-  height: 0;
-  opacity: 0;
-  position: absolute;
-  inset: 0;
-}
-
-.list {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacing-12);
-  margin-top: var(--spacing-32);
-
-  @media (min-width: 768px) {
-    gap: var(--spacing-24);
-  }
-}
-
-.list-item {
-  position: relative;
-  display: flex;
-  gap: var(--spacing-24);
-  list-style: none;
-  font-weight: 300;
-  padding: 0 var(--spacing-16);
-}
-
-.list-item--card {
-  padding: 0;
-}
-
-.skill {
-  font-size: var(--font-size-14);
-
-  @media (min-width: 512px) {
-    font-size: var(--font-size-16);
-  }
-}
-
-.experience-card {
-  display: flex;
-  flex-direction: column;
-  padding: var(--spacing-16);
-  border-color: white;
-  width: 100%;
-
-  @media (min-width: 768px) {
-    flex-direction: row-reverse;
-    padding: var(--spacing-32) var(--spacing-16);
-  }
-}
-
-.experience-card:hover {
-  color: rgb(102 238 255);
-  border-color: rgb(102 238 255);
-}
-
-.experience-card:hover > .experience-card__content {
-  text-decoration: underline;
-}
-
-.experience-card:hover > .experience-card__content > .experience-card__content__icon {
-  transition: transform 0.25s;
-  fill: rgb(102 238 255);
-  transform: translate(var(--spacing-4), calc(-1 * var(--spacing-4)));
-}
-
-.experience-card__content {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--spacing-4);
-  font-weight: 600;
-  width: 100%;
-  fill: white;
-}
-
-.experience-card__content__icon {
-  width: 1rem;
-  margin-left: var(--spacing-8);
-}
-
-.experience-card__content__role {
-  width: 100%;
-}
-
-.experience-card__content__description {
-  margin-top: var(--spacing-12);
-  font-size: var(--font-size-12);
-  opacity: 0.7;
-  font-weight: 300;
-
-  @media (min-width: 512px) {
-    font-size: var(--font-size-14);
-  }
-}
-
-.experience-card__aside {
-  position: absolute;
-  right: var(--spacing-16);
-  top: var(--spacing-40);
-  font-size: var(--font-size-12);
-  width: 100%;
-  max-width: 35%;
-
-  @media (min-width: 768px) {
-    position: relative;
-    inset: 0;
-    font-size: var(--font-size-14);
-  }
-}
-
-.experience-card__aside__date {
-  display: flex;
-  align-items: center;
-  white-space: nowrap;
-  justify-content: flex-end;
-
-  @media (min-width: 768px) {
-    justify-content: flex-start;
-  }
-}
-
-.experience-card__link {
-  position: absolute;
-  inset: 0;
-  width: 100%;
-  height: 100%;
-  border-radius: var(--border-radius-16);
-}
-
-.experience-card__link:hover {
-  background-color: rgb(0 255 255 / 20%);
-}
-
-.experience-card__link__label {
-  opacity: 0;
-}
-
-.project-card {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-
-  @media (min-width: 768px) {
-    flex-direction: row-reverse;
-  }
-}
-
-.project-card__content {
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  padding: var(--spacing-16);
-  fill: white;
-
-  @media (min-width: 768px) {
-    padding: var(--spacing-32) var(--spacing-16) var(--spacing-32) 0;
-  }
-}
-
-.project-card__content:hover {
-  color: rgb(102 238 255);
-  text-decoration: underline;
-  text-underline-position: from-font;
-}
-
-.project-card__content:hover > .project-card__content__header__icon {
-  transition: transform 0.25s;
-  fill: rgb(102 238 255);
-  transform: translate(var(--spacing-4), calc(-1 * var(--spacing-4)));
-}
-
-.project-card__content:hover::before {
-  content: '';
-  position: absolute;
-  inset: 0;
-  width: 100%;
-  height: 100%;
-  border-radius: var(--border-radius-16);
-  background-color: rgb(0 255 255 / 20%);
-  z-index: -1;
-}
-
-.project-card__content__header__icon {
-  width: 1rem;
-  fill: inherit;
-  margin-left: var(--spacing-8);
-}
-
-.project-card__content__description {
-  opacity: 0.7;
-  margin-top: var(--spacing-8);
-  font-size: var(--font-size-12);
-
-  @media (min-width: 512px) {
-    font-size: var(--font-size-14);
-  }
-}
-
-.project-card__content__link {
-  width: 100%;
-  max-width: 80%;
-  height: 100%;
-  position: absolute;
-  inset: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-
-  @media (min-width: 768px) {
-    justify-content: flex-start;
-    left: unset;
-    top: 0;
-    right: 0;
-    max-width: 65%;
-  }
-}
-
-.project-card__aside {
-  width: 100%;
-  max-width: 20%;
-
-  @media (min-width: 768px) {
-    max-width: 35%;
-  }
-}
-
-.project-card__aside__link {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: flex-end;
-  padding: var(--spacing-32) var(--spacing-16);
-
-  @media (min-width: 768px) {
-    justify-content: flex-start;
-  }
-}
-
-.project-card__aside__link:hover::before {
-  content: '';
-  position: absolute;
-  inset: 0;
-  width: 100%;
-  height: 100%;
-  border-radius: var(--border-radius-16);
-  background-color: rgb(0 255 255 / 20%);
-  z-index: -1;
-}
-
-.project-card__aside__link:hover > svg {
-  fill: rgb(102 238 255);
-}
-
-.project-card__aside__link__icon {
-  width: 1.5rem;
-  height: 1.5rem;
-  fill: white;
-
-  @media (min-width: 512px) {
-    width: 2rem;
-    height: 2rem;
-  }
-}
-
-.footer {
-  grid-area: footer;
-  align-self: self-end;
-  padding: 0 var(--spacing-16);
-
-  @media (min-width: 1024px) {
-    position: fixed;
-    bottom: var(--spacing-48);
-  }
-}
-
-.footer__contacts {
-  display: flex;
-  align-items: flex-end;
-  gap: var(--spacing-16);
-}
-
-.footer__contacts__item {
-  list-style-type: none;
-}
-
-.footer__contacts__item__icon {
-  width: 1.5rem;
-  height: 1.5rem;
-  display: flex;
-  align-items: center;
-  fill: white;
-  transition: fill 0.5s;
-
-  @media (min-width: 512px) {
-    width: 2rem;
-    height: 2rem;
-  }
-}
-
-.footer__contacts__item__icon:hover {
-  fill: rgb(102 238 255);
-}
